cloud-function/requirements.txt
google-cloud-logging==3.5.0
google-cloud-firestore==2.14.0
